add a stream parser for fun
